se verifica cuales structuras faltavan

+++ models/message.go
@@ -2,7 +2,7 @@ package models
 
 //Message estructura de los mensajes de respuesta
 type Message struct {
-	Code    int    `json:"code"`
+	Code    uint16 `json:"code"`
 	Data    string `json:"data,omitempty"`
 	Message string `json:"message,omitempty"`
 }
